src/backend/requirements.txt
## The following requirements were added by pip freeze:
algoliasearch==2.6.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.6
cycler==0.11.0
Fiona==1.8.22
fonttools==4.38.0
geopandas==0.12.2
idna==3.4
kiwisolver==1.4.4
matplotlib==3.6.2
munch==2.5.0
networkx==2.8.8
numpy==1.24.0
osmnx==1.2.3
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
pyproj==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
Rtree==1.0.1
Shapely==1.8.5.post1
six==1.16.0
urllib3==1.26.13
